services/users/requirements.txt
fastapi>=0.95.0
uvicorn[standard]
sqlalchemy>=2.0.0
asyncpg
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv
email-validator
pydantic[email]